+++ lib/telnyx/models.rb
@@ -1080,6 +1080,8 @@ module Telnyx
 
   Payment = Telnyx::Models::Payment
 
+  PaymentCreateStoredPaymentTransactionParams = Telnyx::Models::PaymentCreateStoredPaymentTransactionParams
+
   PhoneNumber = Telnyx::Models::PhoneNumber
 
   PhoneNumberBlocks = Telnyx::Models::PhoneNumberBlocks
@@ -1201,9 +1203,17 @@ module Telnyx
 
   PwgAssignedResourcesSummary = Telnyx::Models::PwgAssignedResourcesSummary
 
+  QueueCreateParams = Telnyx::Models::QueueCreateParams
+
+  QueueDeleteParams = Telnyx::Models::QueueDeleteParams
+
+  QueueListParams = Telnyx::Models::QueueListParams
+
   QueueRetrieveParams = Telnyx::Models::QueueRetrieveParams
 
   Queues = Telnyx::Models::Queues
+
+  QueueUpdateParams = Telnyx::Models::QueueUpdateParams
 
   RcsAgent = Telnyx::Models::RcsAgent